New Geometry trait for DSL style component API

+++ diylc-component-library/src/main/groovy/org/diylc/components/boards/PerfBoard.groovy
@@ -25,22 +25,23 @@ public class PerfBoard extends AbstractBoard {
     public static Color COPPER_COLOR = Color.decode("#DA8A67")
 
     public static Size SPACING = new Size(0.1d, SizeUnit.in)
+
     public static Size PAD_SIZE = new Size(0.08d, SizeUnit.in)
+
     public static Size HOLE_SIZE = new Size(0.7d, SizeUnit.mm)
 
     // private Area copperArea
     protected Size spacing = SPACING
+
     protected Color padColor = COPPER_COLOR
 
     @Override
-    public void draw(GraphicsContext graphicsContext, ComponentState componentState,
-                     boolean outlineMode, Project project,
-                     IDrawingObserver drawingObserver) {
+    public void draw(GraphicsContext graphicsContext, ComponentState componentState, boolean outlineMode, Project project, IDrawingObserver drawingObserver) {
         Shape clip = graphicsContext.getClip()
 
         if (checkPointsClipped(clip)
-                && !clip.contains(firstPoint.x, secondPoint.y)
-                && !clip.contains(secondPoint.x, firstPoint.y)) {
+        && !clip.contains(firstPoint.x, secondPoint.y)
+        && !clip.contains(secondPoint.x, firstPoint.y)) {
             return
         }
 
@@ -64,8 +65,8 @@ public class PerfBoard extends AbstractBoard {
                 while (p.x < secondPoint.x - spacing - diameter) {
                     p.@x += spacing
 
-                    graphicsContext.drawFilledOval(p.x - (int) diameter / 2, p.y - (int) diameter / 2, diameter, padColor.darker(), padColor)
-                    graphicsContext.drawFilledOval(p.x - (int) holeDiameter / 2, p.y - (int) holeDiameter / 2, holeDiameter, padColor.darker(), Constants.CANVAS_COLOR)
+                    graphicsContext.drawFilledOval(p.x - diameter / 2, p.y - diameter / 2, diameter, padColor.darker(), padColor)
+                    graphicsContext.drawFilledOval(p.x - holeDiameter / 2, p.y - holeDiameter / 2, holeDiameter, padColor.darker(), Constants.CANVAS_COLOR)
                 }
             }
 
@@ -93,10 +94,10 @@ public class PerfBoard extends AbstractBoard {
 
     @Override
     public void drawIcon(GraphicsContext graphicsContext, int width, int height) {
-        int factor = (int) 32 / width
+        int factor = 32 / width
 
-        graphicsContext.drawFilledRect((int) 2 / factor, width - (int) 4 / factor, BORDER_COLOR, BOARD_COLOR)
-        graphicsContext.drawFilledOval((int) width / 4, (int) width / 2, COPPER_COLOR.darker(), COPPER_COLOR)
-        graphicsContext.drawFilledOval((int) width / 2 - (int) 2 / factor, getClosestOdd(5.0 / factor), COPPER_COLOR.darker(), Constants.CANVAS_COLOR)
+        graphicsContext.drawFilledRect( 2 / factor, width -  4 / factor, BORDER_COLOR, BOARD_COLOR)
+        graphicsContext.drawFilledOval(width / 4,  width / 2, COPPER_COLOR.darker(), COPPER_COLOR)
+        graphicsContext.drawFilledOval(width / 2 - 2 / factor, getClosestOdd(5.0 / factor), COPPER_COLOR.darker(), Constants.CANVAS_COLOR)
     }
 }
